test(passkeys): Add unit tests for passkeys.
Move common test fixtures into a testsupport package.

Configure tests for the passkeys app.
