+++ .coveragerc
@@ -1,5 +1,7 @@
 # https://coverage.readthedocs.io/en/latest/source.html#source
 [run]
-omit=**/tests/**
-  iouproject/asgi.py
-  iouproject/wsgi.py
+omit =
+    **/tests/**
+    testsupport
+    iouproject/asgi.py
+    iouproject/wsgi.py
