+++ conftest.py
@@ -1,4 +1,5 @@
 pytest_plugins = [
     "iou.tests.factories",
     "iou.tests.fixtures",
+    "testsupport.fixtures",
 ]
