+++ testsupport/fixtures.py
@@ -1,0 +1,21 @@
+from django.test import Client
+import pytest
+from django.contrib.auth import get_user_model
+
+User = get_user_model()
+
+@pytest.fixture()
+def test_account_password():
+    return "secret"
+
+
+@pytest.fixture()
+def user(test_account_password):
+    return User.objects.create_user(username="fred", password=test_account_password)
+
+
+@pytest.fixture()
+def client(user: User, test_account_password: str):
+    c = Client()
+    c.login(username=user.username, password=test_account_password)
+    return c
